+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+Version 52:
+
+API Changes:
+
+* auto_fragment is a member of stream
+
+Actions Required:
+
+* Change call sites which use auto_fragment with set_option
+  to call stream::auto_fragment instead.
+
+--------------------------------------------------------------------------------
+
 Version 51
 
 * Fix operator<< for header